fix tests settings template; make_api_call -> recognize_title_via_api; remove unused imports and exceptions; remove process_pubsub_message_v2 and process_pubsub_message_v3.
