+++ src/check_topics_by_upd_time/main.py
@@ -16,7 +16,8 @@ from google.cloud.functions.context import Context
 from google.cloud.storage.blob import Blob
 from retry.api import retry_call
 
-from _dependencies.commons import Topics, get_forum_proxies, publish_to_pubsub, setup_google_logging
+from _dependencies.commons import Topics, get_forum_proxies, setup_google_logging
+from _dependencies.pubsub import publish_to_pubsub
 
 setup_google_logging()
