+++ src/compose_notifications/_utils/notifications_maker.py
@@ -6,8 +6,8 @@ from typing import Any
 import sqlalchemy
 from sqlalchemy.engine.base import Connection
 
-from _dependencies.commons import ChangeType, Topics, get_app_config, publish_to_pubsub
-from _dependencies.misc import notify_admin
+from _dependencies.commons import ChangeType, Topics, get_app_config
+from _dependencies.pubsub import notify_admin, publish_to_pubsub
 
 from .commons import (
     SEARCH_TOPIC_TYPES,
