+++ src/identify_updates_of_topics/_utils/external_api.py
@@ -4,7 +4,7 @@ from geopy.geocoders import Nominatim
 from yandex_geocoder import Client, exceptions
 
 from _dependencies.commons import get_app_config
-from _dependencies.misc import notify_admin
+from _dependencies.pubsub import notify_admin
 
 
 def get_coordinates_from_address_by_osm(address_string: str) -> tuple[float, float]:
